# Reintroduce Shrink.list_spine optimization