--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## NEXT RELEASE
 
-- fix #273 by lowering `ppx_deriving_qcheck`'s `qcheck` dependency to `qcheck-core`
+- Reintroduce the `Shrink.list_spine` fix using an address comparison instead.
+- Fix #273 by lowering `ppx_deriving_qcheck`'s `qcheck` dependency to `qcheck-core`
 
 ## 0.21.1