--- a/src/core/QCheck.ml
+++ b/src/core/QCheck.ml
@@ -760,7 +760,7 @@ module Shrink = struct
     match l with
     | [] -> ()
     | [_] -> yield []
-    | [x;y] -> yield []; yield [x]; yield [y]
+    | [x;y] -> yield []; yield [x]; if x != y then yield [y]
     | _::_ ->
       let len = List.length l in
       let xs,ys = split l ((1 + len) / 2) [] in