This little PR reintroduces a small optimization to `Shrink.list_spine` that had to be rolled back in #280.
The reason was that the polymorphic difference (inequality?) operator `<>` would fail when comparing list elements with function values:
```ocaml
# succ <> pred;;
Exception: Invalid_argument "compare: functional value".
```

Kudos to @shym for suggesting a simple fix: use an address comparison `!=` instead!

On the one hand this
- doesn't raise exceptions as documented by the regression test `test_list_spine_compare`
- eliminates some redundant shrinking candidates as documented by the updated test outputs, and thus saves needless CPU cycles during shrinking

On the other hand, it doesn't identify as many equalities since, e.g., identical strings don't necessarily live at the same address :shrug: